move validation test implementation out of header files
